src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/style-sharing-sibling-shadow-expected.txt

CONSOLE ERROR: line 22: Uncaught TypeError: host.createShadowRoot is not a function
Siblings should only share if their host rules match

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
